requirements.txt
# Core Django dependencies
Django==5.1.7
djangorestframework==3.14.0
gunicorn==21.2.0  # Required for Render

# Database (PostgreSQL support)
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Middleware & CORS
whitenoise==6.6.0  # For serving static files on Render
django-cors-headers==4.3.1

# Image & ML Dependencies
torch==2.1.0
torchvision==0.16.0
Pillow==10.0.1
numpy==1.26.0
dj-database-url
django-cors-headers
dill==0.3.6

